requirements.txt
numpy
opencv-python
scikit-image
tensorflow-macos
tensorflow-metal
matplotlib
pandas